+++ docs/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Fix Jira metadata collector crash on duplicate field names (OSIDB-4920)
 - Fixed double error raise problem in sync_manager.
 
 ## [5.11.0] - 2026-05-28
@@ -21,6 +23,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Exclude streams with no offer from tracker suggestions response (OSIDB-4977)
+- Refresh Flaws in signal to avoid revert issues (OSIDB-4985)
+
+### Fixed
+- Fix check parser replaces spaces with underscores in equality check values (OSIDB-4956)
+
 
 ## [5.10.3] - 2026-05-06
 ### Fixed